refactor: use HTTPFields instead of Header
