+++ Sources/HttpPipeline/Middleware.swift
@@ -19,16 +19,20 @@ public func writeStatus<A>(_ status: Status) -> Middleware<StatusLineOpen, Heade
   }
 }
 
-public func writeHeader<A>(_ header: Response.Header) -> Middleware<HeadersOpen, HeadersOpen, A, A> {
+public func writeHeader<A>(_ header: HTTPField) -> Middleware<HeadersOpen, HeadersOpen, A, A> {
   return pure <<< (\.response.headers %~ { $0 + [header] })
 }
 
-public func writeHeaders<A>(_ headers: [Response.Header]) -> Middleware<HeadersOpen, HeadersOpen, A, A> {
+public func writeHeaders<A>(_ headers: HTTPFields) -> Middleware<HeadersOpen, HeadersOpen, A, A> {
+  return pure <<< (\.response.headers %~ { $0 + headers })
+}
+
+public func writeHeaders<A>(_ headers: [HTTPField]) -> Middleware<HeadersOpen, HeadersOpen, A, A> {
   return pure <<< (\.response.headers %~ { $0 + headers })
 }
 
 public func writeHeader<A>(_ name: HTTPField.Name, _ value: String) -> Middleware<HeadersOpen, HeadersOpen, A, A> {
-  return writeHeader(.init(name, value))
+  return writeHeader(.init(name: name, value: value))
 }
 
 public func closeHeaders<A>(conn: Conn<HeadersOpen, A>) -> IO<Conn<BodyOpen, A>> {
