+++ Sources/HttpPipeline/NIO.swift
@@ -107,7 +107,7 @@ private final class Handler: ChannelInboundHandler {
         let head = HTTPResponseHead(
           version: .init(major: 1, minor: 1),
           status: .init(statusCode: res.status.code),
-          headers: .init(res.headers.map { ($0.name, $0.value) })
+          headers: .init(res.headers.map { ($0.name.rawName, $0.value) })
         )
         context.channel.write(HTTPServerResponsePart.head(head), promise: nil)
