+++ Tests/HttpPipelineHtmlSupportTests/HttpPipelineHtmlSupportTests.swift
@@ -15,7 +15,7 @@ class HttpPipelineHtmlSupportTests: XCTestCase {
     let pipeline = writeStatus(.ok)
       >=> respond(view)
 
-    let conn = connection(from: URLRequest(url: URL(string: "/")!), defaultHeaders: [])
+    let conn = connection(from: URLRequest(url: URL(string: "/")!), defaultHeaders: [:])
     let response = await (conn |> pipeline).performAsync()
 
     XCTAssertEqual(200, response.response.status.code)
