+++ Tests/HttpPipelineTests/HttpPipelineTests.swift
@@ -8,7 +8,7 @@ import Prelude
 import SnapshotTesting
 import XCTest
 
-private let conn = connection(from: URLRequest(url: URL(string: "/")!), defaultHeaders: [])
+private let conn = connection(from: URLRequest(url: URL(string: "/")!), defaultHeaders: [:])
 
 class HttpPipelineTests: XCTestCase {
   override func setUp() {
